requirements.txt
# Discord bot dependencies
discord.py>=2.3.0
python-dotenv>=1.0.0

# Database dependencies
pymongo>=4.6.0
dnspython>=2.4.0

# Oracle database 
oracledb>=2.0.0


